pyproject.toml
[project]
name = "langgraph-agents"
version = "0.1.0"
description = "This is a git repository of AI Agents developed using langgraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.9",
    "langchain-community>=0.3.20",
    "langgraph>=0.6.3",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "ipywidgets>=8.1.5",
    "streamlit>=1.50.0",
    "pandas"
]
